feat(library): rework logging to be user-modifiable
* LogMessage now has a constructor for every type of log message
* in initEnv, check if we should log (i. e. cGopherLogConfig isJust)
  and create or not create a TimedFastLogger accordingly.
  Also Env now contains the wohle initial GopherConfig.
* log uses cGopherLogConfig to determine if and what to log:
  * if isNothing don't log at all
  * if glcTimed is true include timestamp in log messages
  * glcHandler determines if a message is logged (Nothing/Just)
    and formats it (returned LogStr)

This change is preparation for configurable log output in spacecookie.

BREAKING CHANGE:
The type of 'GopherConfig' changed (new field
cGopherLogConfig). This breaks existing code unless
record update syntax was used (as recommended).

Also log output changes, but you shouldn't have relied on the exact
format of the log output in the first place.