fix(config): fail if listen → port is malformed
Previously due to an inadequate use of (<|>), the config parser would
fall back to the regular port option if listen → port was malformed (i.
e. the Parser would return empty). Because that the default option would
be used as the "port" value would be missing in such situations.

We now make sure that (<|>) is only applied to the resulting Maybes and
does not handle any parse failures (empty, mzero, fail …).