fix(Util.Gophermap): require tab for non-info lines
This fixes an incompatibility with bucktooth and pygopherd gophermaps
(although it is documented incorrectly for the latter): Previously we
would parse any line which begins with a valid gopher file type
character as a menu entry and not as a info line.

We now require instead that they contain _at least_ one tab after the
file type and name. Previously you could write:

  0foo

Which would link the file foo. Now the following is required to achieve
the same result:

  0foo^I

This is a breaking change to gophermap parsing which may affect people
using the 'i' file type explicitly to migitate spacecookie being
overzealous when interpreting lines as menu entries. To make up for the
breakage, now info lines won't get misinterpreted anymore.