feat(Util.Gophermap): remove gophermaplineWithoutFileTypeChar
This was a non standard type of line which had no file type character,
but could have an optional path, server and port attached to it and
would be rendered as an info line. This actually makes little sense and
the code was probably only reachable if an invalid gopher file type was
used — which in turn would make the menu entry title prefixed with an
ugly character.

Overall this seems to be an odd relic which was never really in use.
It is probably better to fail to parse in such cases.

BREAKING CHANGE, but will probably affect nobody.