feat(library): use ByteString to represent selectors and paths
In line with the changes made to GopherRequest handlers which now get
the selector as a ByteString instead of a String we also use ByteStrings
for paths and selectors in all other appropriate places.

To elaborate on the reasons for this new philosophy:

* On POSIX, paths are bytes, not strings
* UTF-8 decoding may fail. Instead of handling this in a generic,
  obstrusive way or failing without an user facing error message,
  we should let the library user deal with this issue. Also this issue
  doesn't occur this way if the selectors are directly mapped to a
  posix filesystem, for example.
* Don't force the user to use the inefficient String type, but allow to
  use them the more efficient ByteString or Text if they want to.

For path handling we now use filepath-bytestring instead of filepath
everywhere appropriate. FilePaths are now only used for interacting with
readFile-variants and directory for which no bytestring-based
viable alternative is known at this time.

Some conditional logic has been refactored using boolToMaybe and (<|>).
boolToMaybe is now exposed via Network.Gopher.Util.

BREAKING CHANGES:

* ErrorResponse now also takes a ByteString
* GopherMenuItem now takes an ByteString for describing a path instead
  of a FilePath
* sanitizePath and sanitizeIfNotUrl now are of type
  RawFilePath -> RawFilePath
* gophermapToDirectoryResponse now takes a RawFilePath instead of an
  FilePath
* GophermapFilePath now uses RawFilePaths instead of FilePaths

refactor(server): reflect changes in library