test(server): test not found error message contents
While it is not essential that this text is always the same it should at
least be reproducible and a test failure here may catch an accidental
change to path handling in spacecookie.

If you hit this assertion because you changed the error message feel
free to update the test, this is not an eternal assertion.