chore(cabal): add baseline lower bounds for every dependency
Upper bounds need constant maintenance which is not really worth it as
often the downstream usage doesn't break and if it breaks it breaks at
compile time, so we don't really need upper bounds to tell us that
beforehand.

Ideally I'll notice in time so nobody wastes time with futile
compilation. This should be possible with the cron CI and stackage could
also help if we add spacecookie to that.