fix(Network.Gopher.Util.Gophermap): don't sanitize gophermap paths
The intention of sanitizePath is to eliminate `..` from paths to
prevent path traversal attacks. However, this is unnecessary for
gophermaps for two reasons:

1. A server wouldn't access files based on gophermaps or the menu
   produced from them. The menu would be served to a client which
   would then request it from the server. A malicious selector
   could be produced requested from any gophermap, so the only
   thing that matters is proper sanitization of the selector.

2. In basically all cases, gophermaps come from a trusted source.

Down the line, it would be interesting to support `..` even to link to
the parent directory. This is currently not possible since (Raw)FilePath
won't textually resolve `..` due to the possibility of a path component
being a symlink on a normal unix filesystem.