fix(Network.Spacecookie.Path): more robust makeAbsolute
The previous implementation assumed that a relative path always starts
with "./" or is equal to ".". This isn't true, of course. It was a safe
assumption in the context of spacecookie since `makeAbsolute` is only
ever called on paths produced by

      path = "." </> dropDrive (sanitizePath selector)

The assumption is at risk of being invalidated by refactors, so let's
assume less about the path.