refactor(Network.Spacecookie.FileType): allow singular "." in paths
"." path components are harmless and eliminated by normalisation
anyways. Since we stop caring about them, the dot file check can be
implemented in a much simpler fashion which is done in containsDotFiles
now.

Additionally, directoryResponse needs to be changed since it previously
relied on the fact that checkNoDotFiles _happened_ to eliminate `.` and
`..` from the list returned by getDirectoryContents. We can and should
replace it with listDirectory which doesn't return those entries in the
first place.