refactor(Gopher.Util): replace u*code utils with utf8-string
utf8-string seems to be the most popular of the small encoding packages
(more so than hxt-unicode, utf8-light), so it'll receive attention as
Stackage and GHC move forward. It also provides drop in replacements for
uEncode and uDecode.

An alternative would be using text, but here we'd need to round-trip via
Text for some uses (at least when using the non internal API) which
isn't desireable.

One additional benefit of UTF8.toString is that it's total and won't
crash on invalidly encoded characters which was the case for uDecode.

API BREAKING CHANGE: removed module Network.Gopher.Util